+++ scripts/defaults.js
@@ -8,8 +8,11 @@ export let DEFAULTS = null
 Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
     const groups = GROUP
     Object.values(groups).forEach(group => {
-        group.name = coreModule.api.Utils.i18n(group.name)
-        group.listName = `Group: ${coreModule.api.Utils.i18n(group.listName ?? group.name)}`
+        group.name = group.nameData
+            ? game.i18n.format(group.name, group.nameData)
+            : coreModule.api.Utils.i18n(group.name)
+        group.listName = `Group: ${group.listName ? coreModule.api.Utils.i18n(group.listName) : group.name}`
+        delete group.nameData
     })
     const groupsArray = Object.values(groups)
     DEFAULTS = {
@@ -23,6 +26,14 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
                     { ...groups.damage_adjust, nestId: 'dicepool_damage' },
                     { ...groups.horror_adjust, nestId: 'dicepool_horror' },
                     { ...groups.dicepool_actions, nestId: 'dicepool_actions' }
+                ]
+            },
+            {
+                nestId: 'insight',
+                id: 'insight',
+                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Insight'),
+                groups: [
+                    { ...groups.insight, nestId: 'insight_actions' },
                 ]
             },
             {
@@ -50,11 +61,21 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
                 ]
             },
             {
-                nestId: 'insight',
-                id: 'insight',
-                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Insight'),
+                nestId: 'weapons',
+                id: 'weapons',
+                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Weapons'),
                 groups: [
-                    { ...groups.insight, nestId: 'insight_actions' },
+                    { ...groups.weapons_melee, nestId: 'weapons_melee' },
+                    { ...groups.weapons_ranged, nestId: 'weapons_ranged' },
+                    { ...groups.weapons_other, nestId: 'weapons_other' },
+                ]
+            },
+            {
+                nestId: 'spells',
+                id: 'spells',
+                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Spells'),
+                groups: [
+                    { ...groups.spells, nestId: 'spells_list' },
                 ]
             },
             {
@@ -78,21 +99,16 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
                 ]
             },
             {
-                nestId: 'weapons',
-                id: 'weapons',
-                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Weapons'),
+                nestId: 'knacks',
+                id: 'knacks',
+                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Knacks'),
                 groups: [
-                    { ...groups.weapons_melee, nestId: 'weapons_melee' },
-                    { ...groups.weapons_ranged, nestId: 'weapons_ranged' },
-                    { ...groups.weapons_other, nestId: 'weapons_other' },
-                ]
-            },
-            {
-                nestId: 'spells',
-                id: 'spells',
-                name: coreModule.api.Utils.i18n('ARKHAM_HORROR.LABELS.Spells'),
-                groups: [
-                    { ...groups.spells, nestId: 'spells_list' },
+                    { ...groups.knacks_untiered, nestId: 'knacks_untiered' },
+                    { ...groups.knacks_tier1, nestId: 'knacks_tier1' },
+                    { ...groups.knacks_tier2, nestId: 'knacks_tier2' },
+                    { ...groups.knacks_tier3, nestId: 'knacks_tier3' },
+                    { ...groups.knacks_tier4, nestId: 'knacks_tier4' },
+                    { ...groups.knacks_weaknesses, nestId: 'knacks_weaknesses' },
                 ]
             },
             {
@@ -100,6 +116,7 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
                 id: 'items',
                 name: coreModule.api.Utils.i18n('tokenActionHud.template.items'),
                 groups: [
+                    { ...groups.protective_equipment, nestId: 'items_protective' },
                     { ...groups.useful_items, nestId: 'items_useful' },
                     { ...groups.relics, nestId: 'items_relics' },
                     { ...groups.tomes, nestId: 'items_tomes' },